Macroeconomics
This paper examines the relationship of unemployment rates with other macroeconomic aggregates in Bangladesh over 1991-2019 using robust econometric analyses. It sheds a light on the fact that GDP growth rate, inflation, and foreign direct investment flows have statistically significant impacts on the unemployment rate both in the short-run and long-run. More specifically, the paper documents that unemployment rate, GDP growth rate, inflation rate, and foreign direct investment flows are co-integrated in the long-run at a 5% significance level. Using Vector Error Correction analysis, the paper finds that cointegrated series converge their long-run equilibrium at a speed of 17.24% per annum at a 1% significance level. In the case of the short-run, the study finds that a unit increase in GDP growth rate decreases unemployment by approximately 0.0159 units in the short-run at a 1% statistical significance level. Likewise, a unit increase in the inflation rate will lead to an approximately 0.004 units drop in the unemployment rate at a 10% significance level. Plus, it also observes that a unit in Foreign Direct Investment flows causes 0.005 units decrease in the unemployment rate in the short-run at a 5% significance level.


_________________________________________
Citation (formatted-APA):

 MUKIT, M. M. H., ABDEL-RAZZAQ, A. I., & ISLAM, M. S. (2020). Relationship Between Unemployment and Macroeconomics Aggregates: Evidence from Bangladesh. Journal of Economics and Financial Analysis, 4(2), 45–61. DOI: https://doi.org/10.6084/m9.figshare.13661021.v2. Ashgabat-Turkmenistan, Russian Federation.